+++ app/app.module.ts
@@ -4,13 +4,15 @@ import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
 
-//my-app local
+//my-app componnets local
 import { AppComponent } from './app.component';
 import { UserComponent } from './components/user.component';
+import { AboutComponent } from './components/about.component';
+import { routing } from './app.routing';
 
 @NgModule({
-  imports: [BrowserModule, FormsModule, HttpModule],
-  declarations: [AppComponent, UserComponent],
+  imports: [BrowserModule, FormsModule, HttpModule, routing],
+  declarations: [AppComponent, UserComponent, AboutComponent],
   bootstrap: [AppComponent]
 })
 export class AppModule { }